(chore) dango: extract shared subtle-bytes helper + flatten manifest output
Self-review cleanup:

- Extract u8 / base64ToBytes / ZERO_IV from aes-cbc.ts, hmac.ts, and
  gateway-decrypt.ts into a single _subtle-bytes.ts. All three needed
  the same Uint8Array<ArrayBuffer> workaround for WebCrypto's typing
  and the same raw base64 decoder.
- Flatten the migu episodes pipeline's nested ternary into an assign
  step that picks the source array (datas vs single-episode fallback),
  with the output expression just projecting picked.items.
- Replace the hanjutv danmaku pipeline's fixed \$range(0, 30) with a
  240-iter cap + breakOn that stops on the first empty page. Mirrors
  the upstream's pagination loop without burning fetches on shows that
  end well before the 30-minute mark.
- Drop the duplicate _plaintext field from hanjutv-search.json (the
  test asserts on the decrypted output via the manifest run; the
  duplicate plaintext could drift silently).
- Drop a narration comment from builtin-hanjutv.test.ts.
