+++ packages/dango-manifests/src/__tests__/fixtures/hanjutv-search.json
@@ -1,21 +1,5 @@
 {
-  "_comment": "Encrypted s5 search response. The 'data' field is the base64 ciphertext of the seriesData JSON below, encrypted under AES-128-CBC with no padding. Key/IV derive from md5(uid + ts) then md5(<that> + RESPONSE_SECRET), splitting the resulting 32-char hex into key[0:16] + iv[16:32] (UTF-8 bytes). Verified against the upstream decode path.",
+  "_comment": "Encrypted s5 search response. data = AES-128-CBC ciphertext (no padding) of the seriesList JSON. Key/IV = split(md5(md5(uid + ts) + RESPONSE_SECRET), 16). Regenerate by encrypting the matching plaintext (asserted in the test) under the same scheme.",
   "ts": 1700000000000,
-  "data": "sZSnVwmhGYtBs/2q/QmFNZ2HcuNa7AVEZebJJFo+4zzJQSRVJBXgOwGAqvbeXwY4bzHu+Efj55q/+wBPRSM6NckB/PpdNj3GlPgb8fx4YxN9KSaIRH4gByzO52sAcmggHQNMLx4Y+rd+Ivt+5lB7r7QpF5dofU/D7etGP8pquJUceXEpmdWqFl1GeSymcaxRyEr/UIoq0Yi4t4xM97/40/INl6iB2RjZT/DyVfbt0EqcgN9Z1KXxRZJooZf2h9AGjyBcJX48Pt7dDbmshLdjZnUgJNqA6kjOT5gLOLIRTvPfeoXeOgRwnPb1sD2FcBKCZW7zpvdqJ5kdEB2bgNxrTw==",
-  "_plaintext": {
-    "seriesData": {
-      "seriesList": [
-        {
-          "sid": "hjtv-12345",
-          "name": "示例韩剧",
-          "image": { "thumb": "https://hxqapi.hiyun.tv/asset/12345/poster.jpg" }
-        },
-        {
-          "sid": "hjtv-67890",
-          "name": "第二韩剧",
-          "image": { "thumb": "https://hxqapi.hiyun.tv/asset/67890/poster.jpg" }
-        }
-      ]
-    }
-  }
+  "data": "sZSnVwmhGYtBs/2q/QmFNZ2HcuNa7AVEZebJJFo+4zzJQSRVJBXgOwGAqvbeXwY4bzHu+Efj55q/+wBPRSM6NckB/PpdNj3GlPgb8fx4YxN9KSaIRH4gByzO52sAcmggHQNMLx4Y+rd+Ivt+5lB7r7QpF5dofU/D7etGP8pquJUceXEpmdWqFl1GeSymcaxRyEr/UIoq0Yi4t4xM97/40/INl6iB2RjZT/DyVfbt0EqcgN9Z1KXxRZJooZf2h9AGjyBcJX48Pt7dDbmshLdjZnUgJNqA6kjOT5gLOLIRTvPfeoXeOgRwnPb1sD2FcBKCZW7zpvdqJ5kdEB2bgNxrTw=="
 }
